fix: Se implmento la logica de vehiculos,conductor

+++ src/main/java/org/gersystem/transporte/infrastructure/adapters/rest/dto/ConteoVehiculosDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class ConteoVehiculosDTO {
     private Long conductorId;
     private String nombreConductor;
-    private Long cantidadVehiculos;
+    private Integer cantidadVehiculos;
 }
